Testes de soma, subtração, divisão é multiplicação para portugol (EDITADO)
