+++ testes/portugol-studio/lexador.test.ts
@@ -1,4 +1,4 @@
-import { LexadorPortugolStudio } from '../../fontes/lexador/dialetos';
+import { LexadorPortugolStudio } from "../../fontes/lexador/dialetos";
 
 describe('Lexador (Portugol Studio)', () => {
     describe('mapear()', () => {
@@ -30,6 +30,70 @@ describe('Lexador (Portugol Studio)', () => {
 
                 expect(resultado).toBeTruthy();
                 expect(resultado.simbolos).toHaveLength(9);
+            });
+
+            it('Operação matematica - adicao.', () => {
+                const resultado = lexador.mapear([
+                    'programa',
+                    '{',
+                    '   ', 
+                    '    funcao inicio()',
+                    '    {',
+                    '        esreva(1 + 1)',
+                    '    }',
+                    '}'
+                ], -1);
+
+                expect(resultado).toBeTruthy();
+                expect(resultado.simbolos).toHaveLength(15);
+            });
+
+            it('Operação matematica - subtracao.', () => {
+                const resultado = lexador.mapear([
+                    'programa',
+                    '{',
+                    '   ', 
+                    '    funcao inicio()',
+                    '    {',
+                    '        esreva(1 - 1)',
+                    '    }',
+                    '}'
+                ], -1);
+
+                expect(resultado).toBeTruthy();
+                expect(resultado.simbolos).toHaveLength(15);
+            });
+
+            it('Operação matematica - multiplicacao.', () => {
+                const resultado = lexador.mapear([
+                    'programa',
+                    '{',
+                    '   ', 
+                    '    funcao inicio()',
+                    '    {',
+                    '        esreva(1 * 1)',
+                    '    }',
+                    '}'
+                ], -1);
+
+                expect(resultado).toBeTruthy();
+                expect(resultado.simbolos).toHaveLength(15);
+            });
+
+            it('Operação matematica - divisao.', () => {
+                const resultado = lexador.mapear([
+                    'programa',
+                    '{',
+                    '   ', 
+                    '    funcao inicio()',
+                    '    {',
+                    '        esreva(1 / 1)',
+                    '    }',
+                    '}'
+                ], -1);
+
+                expect(resultado).toBeTruthy();
+                expect(resultado.simbolos).toHaveLength(15);
             });
         });
     });
